Return -ETIMEOUT if vas_paste() timeout
This commit is related with issue #161.
In unit test, if we set 'paste_retries' to 1, using 200 threads doing compress/uncompress,
the vas_paste() is very likely to return '0', which means to retry paste. In this case,
nxu_run_job() returns '0'. nx_submit_job() regards '0' as success. The following
nx_inflate()/nx_deflate() will be in wrong state.

The default 'paste_retires' value is 5000. So normally we will not see vas_paste()
retry failure.