Avoid building unnecessary files during build
Some files for samples/ and test/ were getting built unnecessarily when
calling make.
Furthermore, some of these files depend on the availability of the
libraries, causing errors when calling make with a high number of jobs.

This patch also introduces target bench, which allows to build files
related to benchmarks.