Make separate api lib without the zlib api for some tests
This commit changes the build process to also build a static libnxz
without the zlib interface symbols.

This libray is then use for some of the tests that were meant to compare
the output of libnxz with zlib, so that the zlib function calls actually
call into zlib, ant not libnxz through the zlib interface.