oct: Add output comparison tests
Add an initial implementation of output comparison tests.
These work by generating or downloading different file formats with
different sizes.  These files are later compressed and decompressed and
the final result is compared with the initial one.
The tests also validate if files compressed with the system software are
decompressed with Libnxz and vice-versa.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>