Trust make to generate object files
This recipe was forcing the build of all files unnecessarily.
Furthermore, it isn't really required because GNU make is able to
generate .o files automatically based on their respective .c file
provided that all compiler flags are passed, i.e. in this case CFLAGS.

After doing this, the variable SRCS becomes unused and can be removed.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>