Increase source_sz when resuming decompression after ERR_NX_DATA_LENGTH
The NX GZIP user manual predicts a scenario of insufficient source where
cc=ERR_NX_DATA_LENGTH, sfbt=0xe and subc >= 0.
This patch fixes a corner case of this scenario when the accelerator is
not able to progress because it has insufficient source and source_sz is
not updated.
When that happens, jobs are submitted to the accelerator repeatedly
until loop_cnt reaches loop_max when Z_STREAM_END is returned despite
the lack of progress, causing an application to wrongly conclude the
inflate was complete.

In order to prevent this error from happening, 2 changes are made:

1. Z_STREAM_ERROR is returned when loop_max is reached because an
   application has to abort in that case.
2. Set the last compression ratio to the maximum value in order to
   increase source_sz to its maximum value when the accelerator returns
   insufficient source.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>