Deduplicate code and abort tests in case of unexpected return codes
In order to fix issue #59, many implementations of _test_deflate,
_test_inflate, _test_nx_deflate and _test_nx_inflate would have to be
modified.
Most of these functions are either copy or had a close that could be
shared between most of the tests.  So, this patch move the
implementation of these 4 functions to test_utils.c and reuse them on
most of the tests.
The only exceptions are:
1. deflate/z_finish.c, which has a different implementation of
   _test_nx_deflate;
2. test_stress.c, which has extra code to measure the time of the execution.

The final code still has room for deduplication, but I believe it's an
improvement.

Besides fixing issue #59 and deduplicating code, this patch helps to
reduce the amount of warnings from issue #56.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>