Add env var to control number of threads on test_multithread_stress
Currently the number of threads used in test_multithread_stress is fixed and
maps to the output of 'nproc'. On PowerVM systems with limited NX Gzip credits,
this can cause the test to fail while trying to open more windows than
allowed.

The kernel still doesn't export to user space the amount of credits available,
so we can't automatically scale the test yet. As a temporary workaround, let the
number of threads be configurable through an environment variable. Example:

          $ TEST_NTHREADS=40 make check

will make sure test_multithread_stress runs with a 40 threads.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>