Remove unnecessary checks during configuration processing
GCC 12 (dev) started to complain that some NULL checks were not necessary as the
values could never be NULL. The warnings seem to be correct, so the checks have
been removed.

In find_cfg and nx_dump_cfg the loop iterates over a maximum of
cfg_table->cfg_num elements. As long as cfg_num is at most MAX_CONFIG_ITEM, we
are fine because cfg_table can hold that many elements, and key and val are
pointers to "static" arrays inside struct cfg_item.

In set_cfg, we access key and val of item at index cfg_cnt, which is guarded by
the function to be less than MAX_CONFIG_ITEM, so we are guaranteed to access
initialized data.

Tested with gcc 11.1.1 and gcc 12.0.0 (rev 5f0285418940).

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>