Add inflateSyncPoint
The situation that `inflateSyncPoint` checks for is: no more input and waiting
for beginning of LEN field inside a literal block header.

Context: apparently that is/was useful because some network applications
(e.g. rsync) used a hack to save some bytes on the wire when transporting
deflate streams: if both sender and receiver agree to end every network
packet (i.e. partial deflate stream) with a sync flush (Z_SYNC_FLUSH), every
packet payload will end with bytes 00 00 ff ff (LEN/NLEN for 0-length literal
block), so they can be omitted. The receiver then calls inflateSyncPoint after
decompressing the received packet as an extra error checking mechanism.

zlib's implementation is short and simple, because it only has to check two
things: the current state of inflate's state machine and the number of input
bits available. Both are part of the private stream state.

But since on libnxz the biggest part of the state machine is handled by the NX
engine instead, we need to find a way to infer the state the inflate hardware
was on when it stopped.

In NX Gzip terms that's not as straightforward to check as on zlib. We need
to check for a combination of SFBT, SUBC, FC and also the contents of the
two last bytes (see comments in the code).

Fixes #77

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>