Add test for inflateSyncPoint
This new test checks inflateSyncPoint at the following scenarios:
     1. Before decompression begins
     2. When not at a sync point
     3. At a sync point with different bit offsets

A deflate stream generated from a Z_SYNC_FLUSH (which contains a sync point) is
always byte aligned. So the compressor may add a few padding bits after the
BTYPE for the 0-length literal block to make that happen. To force the check for
all possible amounts of padding bits (0-7), this test splits a large block of
data into smaller chunks, which are passed to deflate one at a time, always with
Z_SYNC_FLUSH, so the generated deflate stream will contain one 0-length literal
block (and thus a sync point) for each chunk.

If we have enough chunks, the probability of having all different padding
amounts is high. The values of NCHUNKS and CHUNKSIZE were chosen empirically,
after checking that all padding amounts were generated on basically every run.

Suggested-by: Bulent Abali <abali@us.ibm.com>
Signed-off-by: Matheus Castanho <msc@linux.ibm.com>