Let user adjust soft_copy_threshold
There is no current way to use a non-default soft_copy_threshold, so the user is
stuck with the value of 1024. This commit lets the user specify a new value
through the config file.

But the user is not completely free to choose any arbitrary value. This commit
also sets min and max values for soft_copy_threshold. The minimum was determined
by checking the minimum value needed by the engine to make progress at all
times. For the maximum, going beyond 8KB-2 (8190) makes the tests fail. This
value could change in the future after further investigation.

Co-authored-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>
Signed-off-by: Matheus Castanho <msc@linux.ibm.com>