Return early when inflate input is lower than soft_copy_threshold
Whenever the application calls inflate with an amount of data lesser than
soft_copy_threshold, that data is copied to fifo_in before it's sent to
the NX GZIP accelerator.
After copying this data, the library can return immediately to the
application in order to receive more data and reducing the amount of
requests sent to accelerator.

Fix issue #84.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>