Add test for deflate with zero input data
This test aims to be a more suitable replacement for samples/zpipe-issue99.c,
formatted as a test case instead of a sample program.

It checks deflate return values for different flush parameters when no input is
given (similar to zpipe-issue99), but also tests 0-length files (issue #76).

The expected return values are based on the output from zlib. Specially for the
empty file tests, the output stream from zlib and libnxz will probably be
different, which is OK. But the return values of the intermediate calls to
deflate should be equivalent.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>