Add configuration cache_threshold
cache_threshold indicates the amount of data that should be cached
from nx_deflate() and nx_inflate() before sending a request to NX.

This new configuration replaces compress_threshold that was used
only in nx_deflate().

After this change, soft_copy_threshold is used only to indicate if the
copy/paste facility should be used or if memcpy() is preferred.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>