Avoid goto when handling buffer states on nx_deflate
For buffer states 0b0000 and 0b1000 we already know that:
   1. used_out is 0
   2. avail_in is 0
   3. used_in is 0

When the code jumps to s1 it will always end up in the same case, executing the
code to handle the stream end, and the function will return from there. So we
can move the code to a new function (nx_stream_end) and call it directly,
returning it's result, avoiding the goto and making the code more readable.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>