oct: Also test with minideflate
minideflate is a command line tool similar to gzip, but doesn't generate
gzip-wrapped output. This adds extra compatibility checks to guarantee that
libnxz is able to handle zlib-wrapper streams correctly. zlib-ng's version is
being used here, but since the code is dynamically linked to zlib, it's capable
of using both vanilla zlib and libnxz.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>