Add support for GNU Autoconf
As build systems become complex it becomes harder to automatically
detect if all the requirements for building the library are installed
and to automatically adapt the build to the requirements of users.

GNU Autoconf helps with this by providing a set of tests for known
requirements as well as easy mechanisms to test for requirements.

This allows to stop testing for stuff in Makefiles and giving more power
to the user that is building to the library.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>