Change prefix of all software zlib functions
Most functions that are "direct" calls to software zlib are prefixed with s_,
but there are also other prefixed with sw_, like sw_zlib_init and sw_zlib_close.
The sw_ prefix is more meaningful and aligns well with the 3-char nx_ prefix
used by many libnxz symbols, so use that everywhere.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>