Add test to guarantee ABI stability
This new test uses abidiff (from libabigail project) to guarantee libnxz's ABI
stability in two ways:
  1. Making sure libnxz provides all zlib symbols
  2. Guaranteeing new symbols are not added/removed by mistake

Of course the first check cannot be complete today, because there are still many
zlib functions missing from libnxz. For this reason, the test uses a list of
suppressed symbols to prevent abidiff from complaining about these for now. As
new functions get added, this list should be updated.

The test relies on 2 files that contain the "canonical" representation for
libnxz and libz as of today. They were generated using the abidw tool, also from
libabigail. They contain a textual representation of the libraries' ABIs, and
should be also updated when the ABIs of these two projects change.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>