Fix deflateSetDictionary and inflateSetDictionary (#15)
The biggest fix needed to make these two functions work was to properly
set history and dictionary sizes used by inflate. Both were being round
up to 16 byte boundary, but that is not necessary. The requirement is
that histLen is a multiple of 16, so we can always use the entire
dictionary and add some padding bytes only to the history, making the
length of history + dictionary a multiple of 16.

Besides that, there are some assorted changed to guarantee both
functions have the same behavior as their zlib counterparts.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>