Add support for GNU Automake
Start using GNU Automake in the root directory of the repository.
Replace Makefile with a Makefile.am.
Move code from config.mk.in to specific files while preparing to remove
it.
Remove code from config.mk.in that became redundant.

Add a new configure parameter that disables the support of the Zlib API
by controlling the value of macro ZLIB_API.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>