Start using GNU Automake in lib/
Rename lib/Makefile as lib/Makefile.am and make the changes required to
use Automake and libtool in order to build the library.
This requires changes in configure.ac as well as in files in other
directories.

Libtool has its own 3-part versioning format (i.e. A.B.C) that forces
changes to the final name of the generated library.  However, we can
still adopt the same numbers and continue progressing the version used
in the library provided that functions are not removed from the ABI.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>