Re-add -std=gnu11
After commit 16439619d82f4, config.mk stopped being used during the
build of the library, removing the inclusion of -std=gnu11 to the
compiler flags.  This caused a regression when building with older
compilers that do not enable the GNU C11 by default.

Fixes this issue by adding a configure check that adds -std=gnu11
whenever detecting a compiler that does not enable it by default.

Considering that config.mk is not used anymore, we can safely remove it.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>