Always use generic crc32 for small lengths (#29)
The generic version is faster than the vector-based one up to length 32,
but as the length grows the difference between the two become larger and
larger, with the latter being much faster.

This was all measured empirically on a Power10 machine. We may need to
revisit the length threshold if we ever change the vector-based
implementation.

Fixes #29

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>