Fix types used by adler32 and crc32 functions (#33)
All declarations and functions calls are using the correct types now,
following the types used by the zlib API. There should be no logical
changes.

Summary of changes:
- nx_crc32 and nx_adler32 use size_t for len parameter
- Add crc32_z and make nx_adler32_z an alias to nx_adler32
- Added const qualifier to buf pointer for all crc32-related functions
- *crc32_combine[64] uses off_t for len2
- Changed buf type to unsigned char * for adler32. Our implementation is
  a copy-n-paste of zlib's, and there it uses type Bytef *, which maps to
  unsigned char *, so no changes in the algorithm

The bug also mentioned possible issues with declaration qualifiers. Even
though we don't have those, it's not an issue because:
- ZEXTERN maps to 'extern' on Linux
- ZEXPORT maps to nothing on Linux
- OF was a helper for function prototype declaration, does not apply here.
- FAR was used to work around some quirks fro MSDOS, it maps to nothing
  on Linux

So for our restricted set of targets (Linux on ppc64le) these are not
necessary.

Fixes #33

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>