Fix build on RHEL7
RHEL7 ships with a very old zlib, which did not have crc32_z/adler32_z,
neither the definition of z_size_t. But nx_adler32 was being declared
with the wrong type for len in nx_zlib.h. It now matches its zlib
counterpart.

The GCC available on RHEL7 does not allow initializing a global value
from the value of another variable. So use a macro instead.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>