Remove ZLIB_API
When the zlib API is removed from the library, the generated shared
object becomes binary incompatible with zlib and with a libnxz that
supports the zlib API.  This scenario should be avoided.

The support for dynamic switching between hardware and software
implementations require the zlib API to always be present.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>