Vary history length based on compression level
Adding history bytes to the NX job adds overhead as the library needs to
copy input data to fifo_in to keep the history consistent and also the
engine needs to copy data to an internal memory to seed itself before
starting job processing. Overhead as high as 50% has been observed for
some files.

For this reason it's best to only use the full 32 kB of history allowed
by NX for the highest compression levels. The reasoning for the level ->
history length mapping added here is as follows:

1. History length should be gradually increased for higher levels to
   provide the user some choices to balance between compression ratio
   and time.

2. level 9 (i.e. the highest) should not be the first level to use the
   entire history. There are other techniques we can still use to get
   better compression that may also increase compression time
   significantly, so the user should be able to choose using as much
   history as possible but without using these other techniques. This
   setting currently translates to level 8.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>