Increase the value for max_vas_reuse_count
While running tests on PowerVM with 1MB files, it has been noticed a
huge delay when processing the 100th stream.
The penalty for reopening the VAS window is still too high for 100
executions and is being increased to 10000 in order to keep its execution less
than 10% of the time spent executing those 10000 streams.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>