Dynamically define number of threads for test
If the TEST_NTHREADS envvar is not provided, try to choose a more
appropriate value depending on the virtualization type. On baremetal we
use a larger number of threads, but on PowerVM we do a rough calculation
of how many credits would be needed and compare that against the number
of available credits.

Also remove some duplicated code in favor of what is available from
test_utils.h

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>