Enforce ERR_NX_TPBC_GT_SPBC in deflate on more cases
After libnxz started to use the history when compressing data,
cc = ERR_NX_TPBC_GT_SPBC was not returned all the times it was
expected. That happens because the NX considers SPBC to be a sum of the
amount of input bytes processed plus the length of the history.

Meanwhile, in order to detect ERR_NX_TPBC_GT_SPBC, the length of the
history should not be included in SPBC.

Fixes issue #161.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>