inflateInit: fallback to software if out of credits
Only when the selector mode is set to AUTO or MIX. Fail otherwise. This
may be a little counterintuitive since the application may not be aware
of libnxz, but we assume that if NX_GZIP_TYPE_SELECTOR is set (or a
config file is being used), it must mean the user (or sysadmin) is aware
of it, so it's fair to fail if set to NX or MIX2.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>