Handle mid-compression window suspension
Currently libnxz is not capable of falling back to zlib mid-compression.
So when we suspect a VAS window has been suspended, our only choice is
to keep trying (in case it becomes active again soon) or try open a new
window. nxu_run_job has been changed to apply this strategy. When paste
failures start to happen, it keeps trying for some time, then it checks
the amount of credits in the system to see if a DLPAR operation
occurred. If so, it tries to open a new window.

It's hard (impossible?) to be sure a window was suspended, so we end up
being pessimistic in cases it could be. To help detect that, new windows
register the amount of total credits on window creation, so we can know
if a credit reduction happened.

To try to help the case when multiple processes have suspended windows
and are trying to open a new one, a simple backoff mechanism was
implemented.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>