Test libnxz behavior when out of credits
libnxz behavior upon lack of credits changes depending on the mode being
used (AUTO, NX, SW, etc). The new test test_exhaust_credits checks that
it behaves as expected for each case.

Before starting the test, it creates several child processes to consume
all credits on the system. For this reason, it can cause failures if
other tests are run in parallel with it. So a secondary test suite was
implemented under serial-tests, that forces all tests to run serially.
This should be only be added to this test suite when strictly necessary,
like in this case.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>