Test behavior on window suspension
This test forces a situation in which 2 processes have their windows
suspended and have to fight over a single extra credit to open a new
window.

Just like test_exhaust_credits it consumes all credits on the system, so
it needs to be part of the serial-tests suite. Also, since this test
forces DLPAR operations on the host, it needs to be run as root to be
able to call drmgr to add/remove virtual processors, so it's skipped if
run without super-user privileges.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>