Reduce the scope for SW usage on length threshold
When compressing/decompressing large amounts of data, the first call to
deflate()/inflate() may not have data enough to pass the threshold,
causing libnxz to fallback to software prematurely.
We can only take that decision if the software indicates that all the
input has been made available, i.e. when flush == Z_FINISH.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>