Remove software-hardware balancing based on ratio
GZIP_MIX was supposed to balance calls between NX and software with an
specific ratio for compression and always use software for
decompression. But this behavior is broken today. The ratio is only
being taken into account on inflate (where it shouldn't matter) and on
deflate the ratio does not factor in.

With the many recent improvements to the AUTO mode, the ratio-based
balancing is not that useful anymore, so let's remove it.

Instead, GZIP_MIX now will always compress with NX and decompress with
software (effectively what GZIP_MIX2 does), which is equivalent to using
a ratio of 1.

Also GZIP_MIX2 since it doesn't make sense anymore (equal to GZIP_MIX).
We're not aware of any current users o GZIP_MIX2, so it's unlikely this
change will break anyone's code.

A positive side-effect of this change is that
nx_config.mode.{inflate,deflate} now can only assume 3 values: AUTO, NX
or SW, simplifying deflate and inflate code.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>