[un]compress: Fallback to software if out of credits
Similar to what was done for deflate and inflate, fallback to software
if window allocation failed due to lack of credits.

nx_compress2 and nx_uncompress2 do much more than just initializing the
stream for [de]compression, but the only situation in which they return
with Z_STREAM_ERROR and errno set to EAGAIN is when window allocation
fails due to lack of credits, so we can check for that.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>