Test [un]compress when system is out of credits
test_exhaust_credits has been improved to also check the behavior of
compress and uncompress when out of credits. Besides adding the function
calls, error return codes needed to be changed to accomodate new return
values while staying below the return value limit imposed by the shell.

Also, the strings used as inputs for compression/decompression were too
small, which would make deflate fallback into software in some
situations, so their size was increased.

Finally, in the situations when deflate is expected to fail, we won't
have a valid compressed stream in compr, so an extra call to compress
was added to make sure we generate the compressed data before starting
the actual test.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>