React better to partition migrations
Partition migrations on PowerVM may take a few minutes, and during that
time libnxz will see consecutive paste failures, so we need to keep
waiting until the migration is complete. So use different timeouts for
waiting for a valid CSB and for a successful paste, giving PowerVM more
time on the paste case to properly handle migrations.

nxu_run_job and nx_wait_for_csb now both return -EAGAIN when a timeout
expires, which describes better the nature of the problem than
ETIMEDOUT.

nx_submit_job has also been changed to remove the nasty exit() call in
case of a timeout. Instead, it will now just return the error up the
stack back to the user. Even though this is not ideal, there's not much
else libnxz can do, since it's hard to know if the job can be safely
reissued or not in such cases.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>