Add internal map data structure
This is a simple hash-table-based map to be used by subsequent commits.
It uses a single dynamically-allocated array to store the entries.
Collisions are treated by simply looking for the next available spot in
the array. The array is also automatically resized if the map starts to
run out of space.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>