Make *Reset functions always restore NX state
For readability, the following paragraphs will only mention
deflate-related function names, but the exact same comments apply to
their inflate counterparts.

Previously if deflate had decided to fallback to software deflateReset
would keep using zlib forever instead of being equivalent to consecutive
calls to deflateEnd + deflateInit, as described by the zlib API.

This is particularly troublesome in the following scenario: application
calls deflateInit once, then calls deflate followed by deflateReset in a
loop. If the first deflate call receives too little data, it will
fallback to software, and subsequent calls to deflateReset won't allow
it ever using NX without explicitly calling deflateEnd + deflateInit.

To avoid this problem now deflateReset always restore both zlib and NX
states, leaving libnxz's internal state loaded in strm, guaranteeing
that future calls to deflate/inflate will be able to choose between
hardware and software once again.  The same is true for inflateReset2.

Some tricks are required to accomplish this. First, deflateReset is
supposed to be an optimization, i.e. faster than calling deflateEnd +
deflateInit, so we should strive for the same here. Before this change,
once we decided to fallback to software, we cleared associated NX state.
Now, instead, we leave it allocated and save an association between it
and the current z_stream in an internal map. This association can be
retrieved later by deflateReset to reuse the already-initialized SW and
NX states.

Care must be taken because in the lack of available credits on PowerVM
we may not allocate the NX state on deflateInit, and fallback straight
to zlib. If later the user calls deflateReset, libnxz will try to
initialize the NX state again. But for that we need initialization
parameters originally passed to deflateInit, which libnxz doesn't have
access to since the zlib state is opaque. For this reason the required
parameters are also saved in the map mentioned before.

The entry in the map and both SW and NX states are cleared by
deflateEnd.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>
Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>