Improve window suspension tests
Previously test_window_suspension and test_shared_window_suspension
tried to reduce the amount of credits by 3, but that's harder to do on
LPARs in Dedicated processor mode, which decrease credits by 20 when 1
CPU is removed. In such cases there were many more processes with
suspended windows than the tests expected, so processes could not
properly recover and the test failed.

These two tests have been rewritten to accomodate that scenario. Now
they calculate the number of processes with suspended windows based on
the number of total and used credits in the system, and react
accordingly, which is enough to work on LPARs in both Shared and
Dedicated processor modes.

Fixes #184

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>