Remove nx_fault_storage_address and related code
This variable is a relic from older code that used a signal handler to
catch page faults and set this variable to trigger touching of such
pages before re-issuing an NX job. libnxz doesn't use such mechanism
anymore because it touches pages ahead of job submission, so faults
shouldn't happen. In the end, codepaths depending on the value of
nx_fault_storage_address are just dead code, and can be safely removed.

Signed-off-by: Matheus Castanho <msc@linux.ibm.com>