Remove unused files from samples/
These files were created in order to help with test efforts or to prove
that a certain strategy would work, but they are not used anymore.
They are confusing users that think these programs should be working.

Remove all of them in order to avoid confusion.

If we ever need to look at them again (e.g. samples/gzip_nxdht.c), we
can get them from the Git history.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>