Stop requiring the presence of the device tree
The device tree provides a lot of information besides the amount of NX
GZIP engines available in a system, making it a sensitive directory to
be shared inside a restrictive container.

Add an alternative mechanism to detect the virtualization environment
when the device tree is not available and enumerate 1 NX device if the
NX GZIP device is available and accessible.

Signed-off-by: Tulio Magno Quites Machado Filho <tuliom@linux.ibm.com>